[Pull Request] German Translation of README.md (#121)
**Is your feature request related to a problem? Please describe.**
The original README.md file is only available in English, which makes it difficult for German-speaking users to understand and use the project.

**Describe the solution you'd like**
I have translated the README.md file into German to better serve German-speaking users. The translation includes all essential information and instructions for using the project.

**Describe alternatives you've considered**
I considered leaving the README.md file in English only, but decided that a German translation would provide better accessibility and usability for German-speaking users.

**Additional context**
The original README.md file has been updated and renamed to README-DE.md to clearly indicate that it is a German translation. The translation has been thoroughly reviewed and checked for accuracy.

I have also considered translating the rest of the documentation that is referenced in the text, however, the content is not available in a easily processable form at this time.